Move function to Impl class
