+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/endofyear/EndOfYearManagerImpl.kt
@@ -3,10 +3,14 @@ package au.com.shiftyjelly.pocketcasts.repositories.endofyear
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.EpisodeManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.PodcastManager
+import au.com.shiftyjelly.pocketcasts.utils.DateUtil
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
+import timber.log.Timber
+import java.time.DateTimeException
+import java.time.LocalDate
 import javax.inject.Inject
 import kotlin.coroutines.CoroutineContext
 
@@ -30,4 +34,20 @@ class EndOfYearManagerImpl @Inject constructor(
         getYearStartAndEndEpochMs(year)?.let {
             episodeManager.findListenedCategories(it.start, it.end)
         } ?: flowOf(emptyList())
+
+    private fun getYearStartAndEndEpochMs(year: Int): YearStartAndEndEpochMs? {
+        var yearStartAndEndEpochMs: YearStartAndEndEpochMs? = null
+        try {
+            val date = LocalDate.of(year, 1, 1).atStartOfDay()
+            val fromEpochTimeInMs = DateUtil.toEpochMillis(date)
+            val toEpochTimeInMs = DateUtil.toEpochMillis(date.plusYears(1))
+            if (fromEpochTimeInMs != null && toEpochTimeInMs != null) {
+                yearStartAndEndEpochMs = YearStartAndEndEpochMs(fromEpochTimeInMs, toEpochTimeInMs)
+            }
+        } catch (e: DateTimeException) {
+            Timber.e(e)
+        }
+        return yearStartAndEndEpochMs
+    }
+    data class YearStartAndEndEpochMs(val start: Long, val end: Long)
 }
